// src/app/core/models/auth.models.ts
export interface User {
    id?: number;
    username?: string;
    password?: string;
    firstName?: string;
    lastName?: string;
    displayName?: string;
    name?: string;
    token?: string;
    email?: string;
    phone?: string;
    avatar?: string;
    location?: string;
    title?: string;
}